refactor(cli): parse-run-config 抽 3 个 sub-routine,主体回归 field-default fanout
把 parse-run-config.ts(355 行)按职责拆出 3 个 helper 到同级 parse-run-config/ 目录:
- judge-models.ts: parseJudgeModelsArg / parseJudgeModelsArgOrExit (--judge-models 串解析)
- samples-discovery.ts: discoverSamplesPath (CLI 未传 --samples 时的发现链)
- variant-resolution.ts: resolveVariantSpecs (--control/--treatment/eval.yaml 三态合并 + dedup)

主体 parseRunConfig 收敛到约 230 行,保留约 100 行的 field-default fanout(每行 cli ?? evalConfig ?? default)。这段刻意不再细拆,否则只是把一长串 ?? 散到多文件,反失「精度链一目了然」可读性。

外部 import surface 完全保留:parseJudgeModelsArg / parseJudgeModelsArgOrExit / parseRunConfig / DEFAULT_REPORTS_DIR / RunConfig / ParseRunConfigResult 都还从 parse-run-config.ts 出口(judge-models 两个符号通过 re-export 维持),test 与 sample.ts / studio.ts / gold/compare.ts / evolve.ts 动态 import 路径全部零改动。

测量学红区零改动:Report schema / judge prompt / observe LLM 复盘 prompt / Bootstrap / α / length-debias toggle 全不动,纯搬代码不改语义。
